Add a test-only mechanism for metering most of the relevant resources per-invocation.
This is an opt-in feature that will reset the budget and take a storage snapshot for every 'logical' invocation (such as `call` or a lifecycle operation). Then when the invocation is done we use the snapshot and current budget to produce the estimate for the resources consumed by the invocation (budget-related, IO-related and rent bumps).

This also provides a rough estimation for the respective fee breakdown given a fee config.