Fix an issue in footprint simulation and add some test coverage for it.
The issue occurs when a non-existent entry gets in a failed contract call that is then handled gracefully via `try_call`. The value won't appear in storage and due to how simulation is implemented it also wasn't included in the result footprint due to that. The fix is to just add the footprint-only entries to the recorded ledgers diffs (as a 'never existed' entry).