Emulate access to custom account instance and Wasm in recording auth. (#1476)
### What

Emulate access to custom account instance and Wasm in recording auth.

The change gracefully handles failures (e.g. allows for the instance not
being present) in order to not spam diagnostics in unit tests that don't
care about the footprint (which makes up for majority of the unit
tests).

### Why

Better emulate the authorization behavior (in some edge cases this might
allow for single-pass simulation of custom accounts).

Resolves https://github.com/stellar/rs-soroban-env/issues/1443

### Known limitations

In theory, we could also try calling `__check_auth` on the contract (if
possible), but that will likely produce confusing diagnostics for most
of the contracts (e.g. signature verification errors).