Add a function to inherit the recording auth settings when switching to recording mode (#1495)
### What

Add a function to inherit the recording auth settings when switching to
recording mode.

### Why

This is necessarily for the test utilities that snapshot the auth
manager and unconditionally switch to recording mode (such as helpers
for deploying the contracts). Previously we just disallowed non-root
auth because that didn't matter, but with the constructors it does
matter.

### Known limitations

N/A