Allow passing the setting for non-root auth to the simulation. (#1503)
### What

Allow passing the setting for non-root auth to the simulation.

This is unfortunately a breaking change and thus it needs to be guarded
by unstable-next-api feature.

### Why

This allows for more control over the simulation logic - while the
default is sufficient 99% of the time, there is still 1% of the cases
where non-root auth should be allowed.

### Known limitations

N/A