Add invocation resource metering to check auth testutil. (#1504)
### What

Add invocation resource metering to check auth testutil.

Also harden the unit tests a bit to ensure that budget matches the
CPU/memory we report in invocation resources and add a test case for a
failed call.

### Why

Fix for https://github.com/stellar/rs-soroban-sdk/issues/1421, still
needs to be picked up by the SDK.

### Known limitations

N/A