Remove check_same_env from EnvBase trait (#1516)
### What
Remove the check_same_env function from the EnvBase trait and move it
into the Host impl.

### Why
Functions in the EnvBase are implemented by Host and Guest. The
check_same_env function is only used by the soroban-sdk. In
https://github.com/stellar/rs-soroban-sdk/pull/1436 the soroban-sdk is
no longer going to be calling the function on the Guest, and only
calling it directly on the Host. So there's no longer a need for it to
be part of this trait/interface.