Add support for muxed addresses in SAC transfer. (#1528)
### What

Add support for muxed addresses in SAC transfer.

This introduces a wrapper for a type that represents an address that is
either a regular AddressObject, or a MuxedAddressObject that we then use
in SAC transfer. We also emit the appropriate events as per CAP-67.

This is currently rebased on
https://github.com/stellar/rs-soroban-env/pull/1525

### Why

CAP-67 implementation.

### Known limitations

N/A