Add support for automatic restoration for the 'recording' mode. (#1523)
### What

Add support for automatic restoration for the 'recording' mode.
This includes support in a few different places:

- e2e_invoke
- invocation metering
- simulation

In order to make them all work consistently, the auto-restoration logic
has been moved directly into recording mode of the storage. I did some
cleanup of the unused `pub` storage methods, as it's no longer feasible
to support these.

The `SnapshotSource` interface now assumes that archived entries should
be returned, thus `SnapshotSourceWithArchive` is removed from the
simulation library. However, we still keep `AutorestoringSnapshotSource`
in order to allow bypassing automatic restoration on the host side (this
allows the users to optionally run the simulation without the
auto-restoration simulation and rely on a standalone restore operation
instead).

### Why

Support for CAP-66 features

### Known limitations

N/A