Rename Host::check_same_env to Host::is_same (#1555)
### What
Rename the host comparison method from check_same_env to is_same,
changing it to return a boolean instead of a Result.

  ### Why
The new name better reflects the method's purpose of being used by the
SDK to compare two Hosts and determine if they in fact are the same
underlying host implementation. The check function returns a result with
the host assuming that there is in fact an error, whilst the is function
returns a bool, merely providing information to the SDK so that the SDK
can then decide in the context of its use whether there is an error or a
change in behaviour is required.

Close #1554 

Related:
-
https://github.com/stellar/rs-soroban-sdk/pull/1467#discussion_r2094183786