Add test utilities to compensate hiding the storage interface. (#1556)
### What

Add test utilities to compensate hiding the storage interface.

Also do a small passing-by cleanup for the storage function names.

### Why

There generally should not be a need to expose the Storage
implementation to the env users; the interface is pretty complex and we
want to be able to change it (as we did e.g. for the error handling).
The test utilities make sure the Storage remains mostly encapsulated in
host, while SDK gets the necessary tools to read/write the arbitrary
ledger entries (which is mostly useful for SAC setup - maybe eventually
this should be moved to host as well, as this is protocol level logic).

### Known limitations

N/A