Scale shadow CPU limit in simulation 10x, shadow memory limit 2x. (#1559)
### What

Scale shadow CPU limit in simulation 10x, shadow memory limit 2x.

This picks up a commit released in the 22.1.4 release of env.

### Why

Shadow limit has never been meant as a hard limit that would cause the
simulation to fail. However, since module cache emulation is now
off-loaded to the shadow budget, it became relatively easy to hit the
shadow budget limit by doing repeated calls to the same contract (which
is now cheap for the actual 'enforcing' mode).

Eventually this may become a setting, but that would be nicer to do with
a breaking change in the simulation API in order to not duplicate the
interfaces.

### Known limitations

N/A