Remove the limit from the event debug info implementation. (#1561)
### What

Remove the limit from the event debug info implementation.

Resolves https://github.com/stellar/rs-soroban-env/issues/1288

### Why

It doesn't seem like this provides any tangible safety benefits, as
events are already charged to budget (or shadow budget) and Debug is
only used for tests or verbose logging in Core. If a consumer is
concerned about the amount of data output, it would make more sense to
limit to overall size of the output buffer.


### Known limitations

N/A