Discount code entry rent by 50% (#1562)
### What

Discount code entry rent by 50%. This is based on this CAP-66 update:
https://github.com/stellar/stellar-protocol/pull/1744

### Why

The idea is to reduce the effective rent cost for the code entries, as
they have in-memory size metered for rent now (instead of just the Wasm
size). 2 is a safe factor that doesn't incentivize storing data in the
Wasm entries, as the in-memory size is always greater than 2x of the
code size.

### Known limitations

N/A