Fix simulation for auto-restored contract code. (#1590)
### What

Fix simulation for auto-restored contract code.

When a contract code entry is auto-restored, it will likely miss the
module cache (if it's been in the hot archive), so we should charge the
recompilation to the budget (in the same way as we already do for the
entry that has been just created via upload).

### Why

Improving simulation accuracy.

### Known limitations

N/A