Use v0 instantiation VM costs for cached Wasms.
That reflects the actual on-chain behavior of the module cache, which always unconditionally charges v0 VM instantiation costs.

Curiously, this change causes the simulated and actual instruction counts to go *down*, which is easy to see in the updated tests. However, the intention of the change is to actually fix simulation for the rare cases where v0 cost are higher than v1 cost.

I've also made a small improvement to the auth simulation accuracy, as it fell outside of our 2% simulation accuracy tolerance window.