Use v0 instantiation VM costs for cached Wasms. (#1599)
### What

That reflects the actual on-chain behavior of the module cache, which
always unconditionally charges v0 VM instantiation costs.

Curiously, this change causes the recording and enforcing mode
instruction counts to go *down* in the e2e tests, which is visible in
the PR. However, the intention of the change is to actually fix
simulation for the rare cases where v0 cost are higher than v1 cost.

Note, that in case of the enforcing mode the change happens due to a fix
that replicates the inter-ledger module cache logic in the test setup.
Specifically, we always unconditionally charge v0 instantiation cost
when instantiating a module from the cache, while previously the test
setup usually used v1 instantiation cost. The simulation change
replicates that behavior in the recording mode.

I've also made a small improvement to the auth simulation accuracy, as
it fell outside of our 2% simulation accuracy tolerance window.

### Why

Improving simulation accuracy.

### Known limitations

N/A